Fix DBus socket deletion on Windows

+++ org.argeo.slc.lib.dbus/src/org/argeo/internal/cms/dbus/CmsDBusImpl.java
@@ -71,7 +71,8 @@ public class CmsDBusImpl implements CmsDBus {
 				log.error("Cannot close embedded DBus daemon", e);
 			}
 			try {
-				Files.delete(dBusDaemonSocket);
+				if (Files.exists(dBusDaemonSocket))
+					Files.delete(dBusDaemonSocket);
 			} catch (IOException e) {
 				log.error("Cannot delete DBus daemon socket " + dBusDaemonSocket, e);
 			}
